requirements.txt
Flask==2.3.3
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
influxdb-client==1.38.0
python-dotenv==1.0.0
urllib3==2.0.7
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
blinker==1.6.2 